First

+++ src/main/java/java/com/shelfApp/controller/BookController.java
@@ -1,10 +1,13 @@
 package java.com.shelfApp.controller;
 
+import java.com.shelfApp.model.Book;
+import java.com.shelfApp.repository.BookRepository;
+import java.com.shelfApp.repository.io.JavaIOBookRepositoryImpl;
 import java.util.List;
 
 public class BookController {
     private BookRepository javaIOBookRepository = new JavaIOBookRepositoryImpl();
-    public Book getById(long id){
+    public Book getById(Integer id){
         return javaIOBookRepository.getById(id);
     }
     public List<Book> getAllBooks(){
@@ -13,7 +16,7 @@ public class BookController {
     public Book save(Book book){
         return javaIOBookRepository.save(book);
     }
-    public void deleteById(Long id){
+    public void deleteById(Integer id){
         javaIOBookRepository.deleteById(id);
     }
 }
